+++ tests/e2e/cases/dynamic_require_static_regexp_escape/in.ts
@@ -1,0 +1,4 @@
+const fromLeading = require("./" + RegExp.escape("foo").replace("\\x66oo", "leading"));
+const fromSyntax = require("./" + RegExp.escape("a+b").replace("\\x61\\+b", "syntax"));
+
+console.log(fromLeading.label, fromSyntax.label);